@@ -12,30 +12,30 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~17.3.9",
+    "@angular/animations": "~18.0.0",
     "@angular/cdk": "~17.3.9",
-    "@angular/common": "~17.3.9",
-    "@angular/compiler": "~17.3.9",
-    "@angular/core": "~17.3.9",
-    "@angular/forms": "~17.3.9",
+    "@angular/common": "~18.0.0",
+    "@angular/compiler": "~18.0.0",
+    "@angular/core": "~18.0.0",
+    "@angular/forms": "~18.0.0",
     "@angular/material": "~17.3.9",
-    "@angular/platform-browser": "~17.3.9",
-    "@angular/platform-browser-dynamic": "~17.3.9",
-    "@angular/router": "~17.3.9",
+    "@angular/platform-browser": "~18.0.0",
+    "@angular/platform-browser-dynamic": "~18.0.0",
+    "@angular/router": "~18.0.0",
     "@auth0/angular-jwt": "~5.1.2",
     "rxjs": "~7.8.0",
     "tslib": "~2.3.0",
     "zone.js": "~0.14.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~17.3.7",
+    "@angular-devkit/build-angular": "~18.0.1",
     "@angular-eslint/builder": "~17.5.2",
     "@angular-eslint/eslint-plugin": "~17.5.2",
     "@angular-eslint/eslint-plugin-template": "~17.5.2",
     "@angular-eslint/schematics": "~17.5.2",
     "@angular-eslint/template-parser": "~17.5.2",
-    "@angular/cli": "~17.3.7",
-    "@angular/compiler-cli": "~17.3.9",
+    "@angular/cli": "~18.0.1",
+    "@angular/compiler-cli": "~18.0.0",
     "@types/jasmine": "~4.3.0",
     "@typescript-eslint/eslint-plugin": "^7.2.0",
     "@typescript-eslint/parser": "^7.2.0",
